Fixes changing animations on some characters
Previously some characters like ROB and Little Mac would not reset their
VBN when starting an animation from the first frame. This was because
their animations have no children, and that was a prerequisite for
recognising the "main" animation. Now it's done by assuming
Runtime.TargetAnimation is the main animation.