Hitbox Angles functionality
Add render settings options, config read and write support for: "renderHitboxAngles" and "hitboxAnglesColor" variables.

Adds a function in VBNViewport that takes a hitbox as input and draws an angle indicator. Functional, but unused.

Adds FacingRestriction parameters to (Extended) Special Hitbox for use in correctly deciding angles in two common cases